+++ Signal-Windows/Models/SignalMessage.cs
@@ -6,12 +6,12 @@ namespace Signal_Windows.Models
 {
     public class SignalMessage
     {
-        public uint Id { get; set; }
+        public ulong Id { get; set; }
         public uint Type { get; set; }
         public uint Status { get; set; }
         public string Content { get; set; }
         public string ThreadID { get; set; }
-        public uint? AuthorId { get; set; }
+        public ulong? AuthorId { get; set; }
         public SignalContact Author { get; set; }
         public uint DeviceId { get; set; }
         public uint Receipts { get; set; }
